requirements.txt
pandas>=1.0.0
plotly==4.4.1
cufflinks==0.17.0
pyaml
discord.py
python-dotenv
tabulate